requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyMySQL==1.1.1
pytest==8.3.5
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlmodel==0.0.24
starlette==0.46.0
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
